create data object for ens profiles
