+++ data/ensProfiles.ts
@@ -1,0 +1,269 @@
+export interface Profile {
+    title: string;
+    ensName: string;
+    displayName?: string;
+    twitterProfile: string;
+}
+
+type EnsProfilesType = {
+    nick: Profile;
+    avsa: Profile;
+    rainbowwallet: Profile;
+    mikedemarais: Profile;
+    inzhoop: Profile;
+    spencecoin: Profile;
+    brantly: Profile;
+    chainlinkgod: Profile;
+    cory: Profile;
+    griff: Profile;
+    coltron: Profile;
+    simona: Profile;
+    ceresstation: Profile;
+    liubenben: Profile;
+    garypalmerjr: Profile;
+    master: Profile;
+    superphiz: Profile;
+    jalil: Profile;
+    bosco: Profile;
+    krypto: Profile;
+    wslyvh: Profile;
+    poap: Profile;
+    fireeyesdao: Profile;
+    spikewatanabe: Profile;
+    alextnetto: Profile;
+    elbagococina: Profile;
+    she256: Profile;
+    limes: Profile;
+    nimi: Profile;
+    lefteris: Profile;
+    mihal: Profile;
+    ethlimo: Profile;
+    gregskril: Profile;
+    taytems: Profile;
+    oneEightFour: Profile;
+    premm: Profile;
+    matoken: Profile;
+    jefflau: Profile;
+    james: Profile
+  };
+
+
+export const EnsProfiles: EnsProfilesType = {
+    nick: {
+        ensName: "nick.eth",
+        displayName: "Nick Johnson",
+        title: "Lead developer of ENS & Ethereum Foundation alum",
+        twitterProfile: "https://twitter.com/nicksdjohnson",
+      },
+      avsa: {
+        ensName: "avsa.eth",
+        displayName: "Alex Van de Sande",
+        title: "Ethereum Foundation Alumni & Co-founder of ENS",
+        twitterProfile: "https://twitter.com/avsa",
+      },
+      rainbowwallet:{
+        ensName: "rainbowwallet.eth",
+        displayName: "Rainbow",
+        title: "The fun, simple, & secure way to explore Web3, NFTs, & Ethereum 🌈",
+        twitterProfile: "https://twitter.com/rainbowdotme",
+      },
+      mikedemarais:{
+        ensName: "mikedemarais.eth",
+        displayName: "Mike Demarais",
+        title: "Co-founder of Rainbow Wallet",
+        twitterProfile: "https://twitter.com/mikedemarais",
+      },
+      inzhoop:{
+        ensName: "inzhoop.eth",
+        title: "Independent startup working on smart mobile wallet for Ethereum",
+        twitterProfile: "https://twitter.com/inzhoop",
+      },
+      spencecoin:{
+        ensName: "spencecoin.eth",
+        displayName: "Jordan Spence",
+        title: "Director of Marketing and Strategic Initiatives at Metamask and Consensys",
+        twitterProfile: "https://twitter.com/spencecoin",
+      },
+      brantly:{
+        ensName: "brantly.eth",
+        displayName: "Brantly Millegan",
+        title: "Building Ethereum Follow Protocol & Creator of Sign-In w/ Ethereum",
+        twitterProfile: "https://twitter.com/BrantlyMillegan",
+      },
+      chainlinkgod:{
+        ensName: "chainlinkgod.eth",
+        title: "Chainlink Community Ambassador",
+        twitterProfile: "https://twitter.com/ChainLinkGod",
+      },
+      cory:{
+        ensName: "cory.eth",
+        displayName: "Cory Gabrielsen",
+        title: "Founder OpenAvatar & Co-founder of Ethereum Follow Protocol",
+        twitterProfile: "https://twitter.com/cory_eth",
+      },
+      poap:{
+        ensName: "poap.eth",
+        displayName: "Proof of Attendance Protocol",
+        title: "Bookmarks for your life",
+        twitterProfile: "https://twitter.com/poapxyz",
+      },
+      griff:{
+        ensName: "griff.eth",
+        displayName: "Griff Green",
+        title: "Cofounder Giveth, CommonsStack, GeneralMagic, Dappnode & DECENTRAL",
+        twitterProfile: "https://twitter.com/thegrifft",
+      },
+      coltron:{
+        ensName: "coltron.eth",
+        title: "ENS Public Goods Steward",
+        twitterProfile: "https://twitter.com/Coltron_eth",
+      },
+      simona:{
+        ensName: "simona.eth",
+        title: "ENS Public Goods Steward, Governance at DELV & Optimism",
+        twitterProfile: "https://twitter.com/Sim_Pop",
+      },
+      ceresstation:{
+        ensName: "ceresstation.eth",
+        displayName: "Scott Moore",
+        title: "Co-founder Gitcoin & Kernel",
+        twitterProfile: "https://twitter.com/notscottmoore",
+      },
+      liubenben:{
+        ensName: "liubenben.eth",
+        title: "Building ensuser.com for the Chinese ENS Community",
+        twitterProfile: "https://twitter.com/forlbb",
+      },
+  
+      garypalmerjr:{
+        ensName: "garypalmerjr.eth",
+        title: "ENS Advocate and Founder of Web3Domains",
+        twitterProfile: "https://twitter.com/garypalmerjr",
+      },
+      master:{
+        ensName: "master.eth",
+        title: "ENS Advocate",
+        twitterProfile: "https://twitter.com/seekmine",
+      },
+      superphiz:{
+        ensName: "superphiz.eth",
+        title: "Ethereum Decentralized Staking Advocate",
+        twitterProfile: "https://twitter.com/superphiz",
+      },
+      jalil:{
+        ensName: "jalil.eth",
+        displayName: "Jalil Wahdatehagh",
+        title: "Intangible things at Visualize Value.",
+        twitterProfile: "https://twitter.com/jalil_eth",
+      },
+      bosco:{
+        ensName: "bosco.eth",
+        title: "Love Freedom and Public Goods 💚",
+        twitterProfile: "https://twitter.com/amboscoboinik",
+      },
+      krypto:{
+        ensName: "krypto.eth",
+        title: "Poland ENS Master",
+        twitterProfile: "https://twitter.com/CryptoDodoPL",
+      },
+      wslyvh:{
+        ensName: "wslyvh.eth",
+        displayName: "Wesley",
+        title: "Events at Ethereum Foundation",
+        twitterProfile: "https://twitter.com/wslyvh",
+      },
+      premm:{
+        ensName: "premm.eth",
+        displayName: "Prem Makeig",
+        title: "Founder at Unruggable Labs, ENS Fellow",
+        twitterProfile: "https://twitter.com/nxt3d",
+      },
+      fireeyesdao:{
+        ensName: "fireeyesdao.eth",
+        displayName: "Fire Eyes DAO", 
+        title: "Delegate for Optimism, ENS, GitCoin, Rocket Pool, & Safe",
+        twitterProfile: "https://twitter.com/fireeyesgov",
+      },
+      spikewatanabe:{
+        ensName: "spikewatanabe.eth",
+        title: "ENS Delegate & experienced investment banker",
+        twitterProfile: "https://twitter.com/spikewatanabe",
+      },
+      alextnetto:{
+        ensName: "alextnetto.eth",
+        displayName: "Alex Netto",
+        title: "Co-founder of Blockful, building public goods for DAOs",
+        twitterProfile: "https://twitter.com/alextnetto",
+      },
+      elbagococina:{
+        ensName: "elbagococina.eth",
+        title: "Co-founder Karpatkey, Core Treasury for GnosisDAO, Balancer, ENS, CoWSwap, and Lido",
+        twitterProfile: "https://twitter.com/elbagococina",
+      },
+      she256:{
+        ensName: "she256.eth",
+        title: "Nonprofit dedicated to increasing diversity in the blockchain space",
+        twitterProfile: "https://twitter.com/she_256",
+      },
+  
+      limes:{
+        ensName: "limes.eth",
+        title: "ENS DAO Steward & Secretary - Growth at Layer3",
+        twitterProfile: "https://twitter.com/limes_eth",
+      },
+      nimi:{
+        ensName: "nimi.eth",
+        title: "Your personal web3 page",
+        twitterProfile: "https://twitter.com/0xNimi",
+      },
+      lefteris:{
+        ensName: "lefteris.eth",
+        displayName: "Lefteris Karapetsas",
+        title: "Founder of Rotki, the portfolio tracker that protects your privacy",
+        twitterProfile: "https://twitter.com/LefterisJP",
+      },
+      mihal:{
+        ensName: "mihal.eth",
+        displayName: "David Mihal",
+        title: "Blockchain Engineer",
+        twitterProfile: "https://twitter.com/dmihal",
+      },
+      ethlimo:{
+        ensName: "ethlimo.eth",
+        title: "Privacy-preserving ENS gateway",
+        twitterProfile: "https://twitter.com/eth_limo",
+      },
+      gregskril:{
+        ensName: "gregskril.eth",
+        displayName: "Greg Skriloff",
+        title: "Developer at ENS Labs",
+        twitterProfile: "https://twitter.com/gregskril",
+      },
+      taytems:{
+        ensName: "taytems.eth",
+        title: "Developer at ENS Labs",
+        twitterProfile: "https://twitter.com/taytemss",
+      },
+      oneEightFour:{
+        ensName: "184.eth",
+        title: "Support at ENS Labs & ENS Ecosystem Steward",
+        twitterProfile: "https://twitter.com/184eth",
+      },
+      matoken:{
+        ensName: "matoken.eth",
+        displayName: "Makoto Inoue",
+        title: "Developer at ENS Labs",
+        twitterProfile: "https://twitter.com/makoto_inoue",
+      },
+      jefflau:{
+        ensName: "jefflau.eth",
+        displayName: "Jeff Lau",
+        title: "Core Team Leader - ENS Labs",
+        twitterProfile: "https://twitter.com/_jefflau",
+      },
+      james: {
+        title: "DAO Governance",
+        ensName: "james.eth",
+        twitterProfile: "https://twitter.com/blockchainjames",
+      }
+  };
